add tree renaming
