+++ RadialGM.pro.shared
@@ -30,7 +30,7 @@
             <value type="bool" key="EditorConfiguration.MouseNavigation">true</value>
             <value type="int" key="EditorConfiguration.PaddingMode">1</value>
             <value type="bool" key="EditorConfiguration.ScrollWheelZooming">true</value>
-            <value type="bool" key="EditorConfiguration.ShowMargin">false</value>
+            <value type="bool" key="EditorConfiguration.ShowMargin">true</value>
             <value type="int" key="EditorConfiguration.SmartBackspaceBehavior">0</value>
             <value type="bool" key="EditorConfiguration.SmartSelectionChanging">true</value>
             <value type="bool" key="EditorConfiguration.SpacesForTabs">true</value>
